video list elements now have scroll, better video reproducer page structure
