+++ src/components/VideoReproducerPage/VideoListElements/Thumbnail.jsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Thumbnail = styled.img`
+  margin-right: 20px;
+  width: 150px;
+  @media (min-width: 450px) {
+    width: 250px;
+  }
+`;
+
+export default Thumbnail;
