+++ src/components/VideoReproducerPage/VideoListElements/VideoListElement.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const VideoListElement = styled.div`
+  padding: 0.2rem;
+  background-color: white;
+  border-bottom: rgba(216, 198, 207, 0.5) 2px solid;
+  &:hover {
+    background-color: rgb(241, 217, 217);
+  }
+`;
+
+export default VideoListElement;
